Renamed o-table-columns-filter-columns file to o-table-columns-filter-column
